+++ Soggins/Cards/BlastFromThePastCard.cs
@@ -61,9 +61,10 @@ public sealed class BlastFromThePastCard : Card, IRegisterableCard, IFrogproofCa
 			actions.Add(Instance.KokoroApi.ConditionalActions.Make(
 				Instance.KokoroApi.ConditionalActions.Equation(
 					Instance.KokoroApi.ConditionalActions.Status((Status)Instance.SmugStatus.Id!.Value),
-					IKokoroApi.IConditionalActionApi.EquationOperator.LessThanOrEqual,
+					Instance.Api.GetMinSmug(s.ship) == -3 ? IKokoroApi.IConditionalActionApi.EquationOperator.Equal : IKokoroApi.IConditionalActionApi.EquationOperator.LessThanOrEqual,
 					Instance.KokoroApi.ConditionalActions.Constant(-3),
-					hideOperator: -3 == Instance.Api.GetMinSmug(s.ship)
+					IKokoroApi.IConditionalActionApi.EquationStyle.State,
+					hideOperator: Instance.Api.GetMinSmug(s.ship) == -3
 				),
 				new AStatus
 				{