+++ Soggins/ExternalAPIs/Kokoro/ComplexActions.cs
@@ -15,16 +15,22 @@ public partial interface IKokoroApi
 		IIntExpression ScalarMultiplier(IIntExpression expression, int scalar);
 		IBoolExpression HasStatus(Status status, bool targetPlayer = true, bool countNegative = false);
 		IIntExpression Status(Status status, bool targetPlayer = true);
-		IBoolExpression Equation(IIntExpression lhs, EquationOperator @operator, IIntExpression rhs, bool hideOperator = false);
+		IBoolExpression Equation(IIntExpression lhs, EquationOperator @operator, IIntExpression rhs, EquationStyle style, bool hideOperator = false);
 
 		public enum EquationOperator
 		{
 			Equal, NotEqual, GreaterThan, LessThan, GreaterThanOrEqual, LessThanOrEqual
 		}
 
+		public enum EquationStyle
+		{
+			Formal, State, Possession, PossessionComparison
+		}
+
 		public interface IExpression
 		{
 			void Render(G g, ref Vec position, bool isDisabled, bool dontRender);
+			string GetTooltipDescription(State state, Combat? combat);
 			List<Tooltip> GetTooltips(State state, Combat? combat) => new();
 		}